+++ test/tiny_random.jl
@@ -5,6 +5,7 @@ using DifferentialRiccatiEquations
 using LinearAlgebra, SparseArrays, UnPack
 
 using DifferentialRiccatiEquations.Stuff: delta
+using DifferentialRiccatiEquations.Shifts: Cyclic, Heuristic
 
 n = 50
 g = 4
@@ -21,11 +22,27 @@ function test_ale(E, A, g=g)
     X_adi = solve(prob, ADI())
     X_ref = solve(prob, BartelsStewart())
     X_bad = solve(prob, Kronecker())
+    X_gmres = solve(prob, GMRES(; maxiters=5, reltol=1e-8))
+    X_fgmres = solve(prob, GMRES(;
+        maxiters = 3,
+        maxrestarts = 0,
+        reltol = 1e-10, # default can not be reached
+        preconditioner = ADI(;
+            maxiters = 10,
+            shifts = Cyclic(Heuristic(10, 10, 10)),
+            compression_interval = 20, # only compress final solution
+            warn_convergence = false,
+        ),
+    ))
     @test norm(residual(prob, X_ref)) / res0 < 1e-10
     @test norm(residual(prob, X_adi)) / res0 < 1e-10
     @test norm(residual(prob, X_bad)) / res0 < 0.02
+    @test norm(residual(prob, X_gmres)) / res0 < 1e-8
+    @test norm(residual(prob, X_fgmres)) / res0 < 1e-10
 
     @test delta(Matrix(X_adi), X_ref) < 1e-10
+    @test delta(Matrix(X_gmres), X_ref) < 1e-8
+    @test delta(Matrix(X_fgmres), X_ref) < 1e-10
     @test delta(X_bad, X_ref) < 0.02
 
     solver = init(prob, ADI())